split osArena
