+++ source/rvl/os/osAlloc.h
@@ -20,32 +20,6 @@ UNKNOWN_FUNCTION(set_CurrHeap);
 UNKNOWN_FUNCTION(OSInitAlloc);
 // PAL: 0x801a1038..0x801a10a4
 UNKNOWN_FUNCTION(OSCreateHeap);
-// PAL: 0x801a10a4..0x801a10ac
-UNKNOWN_FUNCTION(OSGetMEM1ArenaHi);
-// PAL: 0x801a10ac..0x801a10b4
-UNKNOWN_FUNCTION(OSGetMEM2ArenaHi);
-// PAL: 0x801a10b4..0x801a10bc
-UNKNOWN_FUNCTION(OSGetArenaHi);
-// PAL: 0x801a10bc..0x801a10c4
-UNKNOWN_FUNCTION(OSGetMEM1ArenaLo);
-// PAL: 0x801a10c4..0x801a10cc
-UNKNOWN_FUNCTION(OSGetMEM2ArenaLo);
-// PAL: 0x801a10cc..0x801a10d4
-UNKNOWN_FUNCTION(OSGetArenaLo);
-// PAL: 0x801a10d4..0x801a10dc
-UNKNOWN_FUNCTION(OSSetArenaHi_0);
-// PAL: 0x801a10dc..0x801a10e4
-UNKNOWN_FUNCTION(OSSetMEM2ArenaHi);
-// PAL: 0x801a10e4..0x801a10ec
-UNKNOWN_FUNCTION(OSSetArenaHi);
-// PAL: 0x801a10ec..0x801a10f4
-UNKNOWN_FUNCTION(OSSetArenaLo_0);
-// PAL: 0x801a10f4..0x801a10fc
-UNKNOWN_FUNCTION(OSSetMEM2ArenaLo);
-// PAL: 0x801a10fc..0x801a1104
-UNKNOWN_FUNCTION(OSSetArenaLo);
-// PAL: 0x801a1104..0x801a1138
-UNKNOWN_FUNCTION(OSAllocFromMEM1ArenaLo);
 
 #ifdef __cplusplus
 }
